Make the message URL in the thread list "read"
This is a cosmetic nicety that makes it clear which message was used to get to
the thread. Same as if you'd manually clicked the message, then clicked "Full
thread".

The way to accomplish this turns out to be stuffing the target URL into the
browser's history and letting it do its job.

Trying to manipulate the styling of the anchor a) didn't work for me for the
naive approach, b) will only ever be a simulation of the desired effect, and c)
will be brittle because it's a simulation.

Stuffing the browser's history is easy enough, once the restriction that the
content script can't access the history API is known about and handled.