Move the history stuffing to after the page loads
Doing the stuffing during the onClick handler is too soon; the message can
become visited before the page loads. Since we're loading a thread which can be
a lot of database queries this could occur when the page never loads at all.

The content script for the thread page handles the history stuffing instead.