Improve URL marking-as-visited performance
Marking as visited is accomplished by stuffing the target URL into the browser
history and letting the browser do its job. But when to stuff?

Stuffing during the thread-link onclick handler is too early; a long thread with
a lot of backend database accesses may fail to load but the message URl will
still be in the history and "visited" even though it wasn't.

Stuffing during a standard content script is too late; there is a
visible-to-the-user delay where the link is styled as unvisited, then visited.

The solution is to stuff at the start of document load, so we know the URL is
going to be at least partially successfully visited, but it happens before the
rendering (I think) so it looks instantaneous to the user. This should replicate
exactly what the browser does natively when clicking a message link, or at
least close enough to native behaviour that it looks native to the user.

The way to accomplish this is to refactor the history stuffing into its own
content script with 'run_at': 'document_start' (normal default is
document_idle).

This works but it turns out that content scripts with the same 'matches' array
are all injected into one shared "isolated world" context, and they can conflict
with each other! I've worked around this for now by enclosing each of the
conflicting scripts in a new {} block. I'm not sure if this is the canonical
solution, or if there is something better. IIRC GreaseMonkey scripts were always
recommended to be written as a
(function(...){...})();
block for cross-script compatibility. Maybe that's what's required here?